refactor: Reorganize localized resource implementaion
