+++ swift-paperless/Localization/Localizable.xcstrings
@@ -3588,6 +3588,17 @@
         }
       }
     },
+    "groups" : {
+      "extractionState" : "manual",
+      "localizations" : {
+        "en" : {
+          "stringUnit" : {
+            "state" : "translated",
+            "value" : "Groups"
+          }
+        }
+      }
+    },
     "importDocument" : {
       "comment" : "Document add menu item for adding files",
       "extractionState" : "manual",
@@ -9378,6 +9389,17 @@
           "stringUnit" : {
             "state" : "translated",
             "value" : "Użytkownik"
+          }
+        }
+      }
+    },
+    "users" : {
+      "extractionState" : "manual",
+      "localizations" : {
+        "en" : {
+          "stringUnit" : {
+            "state" : "translated",
+            "value" : "Users"
           }
         }
       }
